Defalt val. of hpa.metrics to the correct type
This sets the hpa.metrics key to its correct default value. Which is an array and not a dict. Having it as a dict. gives issues as seen in #79. Which I hope to solve with this PR.

See: https://kubernetes.io/docs/reference/kubernetes-api/workload-resources/horizontal-pod-autoscaler-v2/#HorizontalPodAutoscalerSpec for the official ref. docs on HPA and the `metrics` key type specifically.

Thank you.

Signed-off-by: Lars Bengtsson <bingchonglars@gmail.com>
Signed-off-by: Lars Bengtsson <larssb@users.noreply.github.com>