fix: improve default readiness probe config for shutdown
On shutdown coredns will wait what is configured for lameduck and
continue handling connections. At the same time it will fail readiness
probes.

In order to give readiness checks a chance to remove the instance from
the service we need to lower the failure threshold and the interval.

This is to avoid failing DNS requests in a busy cluster when coredns is
being scaled down.

Signed-off-by: Heiko Voigt <heiko.voigt@jimdo.com>